added target 6D pose to pickup object

+++ include/ba_types.h
@@ -4,30 +4,18 @@
 #pragma region includes
 
 #include <cmath>
-// #include <pcl/ModelCoefficients.h>
-// #include <pcl/io/pcd_io.h>
-// #include <pcl/point_types.h>
-// #include <pcl/sample_consensus/method_types.h>
-// #include <pcl/sample_consensus/model_types.h>
-// #include <pcl/sample_consensus/sac_model_sphere.h>
-// #include <pcl/segmentation/sac_segmentation.h>
-// #include <pcl/filters/extract_indices.h>
-// #include <pcl/filters/passthrough.h>
 
 #pragma endregion
-
-// typedef pcl::PointCloud<pcl::PointXYZ>::Ptr BAPointCloudPtr;
-// typedef pcl::PointCloud<pcl::PointXYZ> BAPointCloud;
 
 /**
  * @brief Struct to store tomato coordinates
  * 
  */
 struct TomatoCoordinates{
-    float x;
-    float y;
-    float z;
-    float r;
+    double x;
+    double y;
+    double z;
+    double r;
     bool is_picked;
     inline bool operator==(const TomatoCoordinates t1){
         return (pow(x-t1.x,2)+pow(y-t1.y,2)+pow(z-t1.z,2))<std::max(r, t1.r)+0.01;
@@ -39,9 +27,9 @@ struct TomatoCoordinates{
 //  * 
 //  */
 // struct PathLocation {
-//     float x;
-//     float y;
-//     float orientation;
+//     double x;
+//     double y;
+//     double orientation;
 // };
 
 // /**
