+++ src/helper.cpp
@@ -4,11 +4,11 @@
  * @brief Converts an angle in degrees into radians
  *
  * @param degrees angle to convert in degrees
- * @return float converted angle in radians
+ * @return double converted angle in radians
  */
-float ba_helper::ConvertDegreesToRadians(float degrees)
+double ba_helper::ConvertDegreesToRadians(double degrees)
 {
-    float radians = degrees / 180 * M_PI;
+    double radians = degrees / 180 * M_PI;
     return radians;
 }
 
@@ -16,11 +16,11 @@ float ba_helper::ConvertDegreesToRadians(float degrees)
  * @brief Converts a quaternion into an angle, based on the function: Theta = 2*atan2(sqrt(x^2+y^2+z^2), w)
  *
  * @param quat The quaternion to transform
- * @return float The angle in radians
+ * @return double The angle in radians
  */
-float ba_helper::ConvertQuaternionToAngle(geometry_msgs::msg::Quaternion quat)
+double ba_helper::ConvertQuaternionToAngle(geometry_msgs::msg::Quaternion quat)
 {
-    float axis = 0;
+    double axis = 0;
     axis += quat.x * quat.x;
     axis += quat.y * quat.y;
     axis += quat.z * quat.z;
@@ -32,11 +32,11 @@ float ba_helper::ConvertQuaternionToAngle(geometry_msgs::msg::Quaternion quat)
  * @brief linearisation for gripper command --> calculates msg payload to be published to gripper controller from airgap [m] you want the gripper to open
  *
  * @param airgap Desired opening of the gripper in [m]
- * @return float value to publish to gripper controller for danfoa 140 2f robotiq
+ * @return double value to publish to gripper controller for danfoa 140 2f robotiq
  */
-float ba_helper::CalculateGripperParameterFromDesiredAirgap(float airgap)
+double ba_helper::CalculateGripperParameterFromDesiredAirgap(double airgap)
 {
-    float result = -5.40444426 * airgap + 0.71065;
+    double result = -5.40444426 * airgap + 0.71065;
     if (result > 0.7)
     {
         result = 0.7;
@@ -72,20 +72,15 @@ geometry_msgs::msg::Point ba_helper::GetCurrentArmbaseLocation(void)
  * @return geometry_msgs::msg::PoseStamped The current position in the map frame
  */
 geometry_msgs::msg::PoseStamped ba_helper::GetCurrentPoseInMapFrame(const char source_frame[],
-                                                               float x, float y, float z,
-                                                               float roll, float pitch, float yaw)
+                                                               double x, double y, double z,
+                                                               double roll, double pitch, double yaw)
 {
-    // tf2_ros::TransformListener listener;
-
-    // listener.waitForTransform(MAP_FRAME, source_frame, ros::Time(0), ros::Duration(3.0));
     geometry_msgs::msg::PoseStamped current_position_base_frame;
     geometry_msgs::msg::PoseStamped current_position_map_frame;
 
     current_position_base_frame.pose.position.x = x;
     current_position_base_frame.pose.position.y = y;
     current_position_base_frame.pose.position.z = z;
-
-    // current_position_base_frame.pose.orientation = tf::createQuaternionMsgFromRollPitchYaw(roll, pitch, yaw);
 
     tf2::Quaternion tf2_quat;
     tf2_quat.setRPY(roll, pitch, yaw);
@@ -94,10 +89,8 @@ geometry_msgs::msg::PoseStamped ba_helper::GetCurrentPoseInMapFrame(const char s
     current_position_base_frame.pose.orientation = msg_quat;
 
     current_position_base_frame.header.frame_id = source_frame;
-    // current_position_base_frame.header.stamp = ros::Time();
-    // listener.transformPose(MAP_FRAME, current_position_base_frame, current_position_map_frame);
+
     current_position_map_frame.header.frame_id = MAP_FRAME;
-    // current_position_map_frame.header.stamp = ros::Time();
 
     return current_position_map_frame;
 }
